requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
pypdf
transformers
tk>=8.6